+++ src/components/RecipeModal.jsx
@@ -1,10 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import Modal from 'react-responsive-modal';
+import firebase from '../firebase.js';
 
 export class RecipeModal extends React.Component {
    state = {
       open: false,
+      title: '',
+      image: '',
+      link: '',
    };
 
    onOpenModal = () => {
@@ -15,13 +19,70 @@ export class RecipeModal extends React.Component {
       this.setState({ open: false });
    };
 
+   handleChange = event => {
+      this.setState({
+         [event.target.name]: event.target.value,
+      });
+   };
+
+   handleSubmit = event => {
+      event.preventDefault();
+      const recipesRef = firebase.database().ref('recipes');
+      const recipe = {
+         title: this.state.title,
+         image: this.state.image,
+         link: this.state.link,
+      };
+      recipesRef.push(recipe);
+      this.setState({
+         title: '',
+         image: '',
+         link: '',
+      });
+   };
+
    render() {
       const { open } = this.state;
       return (
          <div>
             <button onClick={this.onOpenModal}>Open modal</button>
             <Modal open={open} onClose={this.onCloseModal} center>
-               <h2>Simple centered modal</h2>
+               <form onSubmit={this.handleSubmit} action="">
+                  <p>
+                     <label htmlFor="title">
+                        Title
+                        <input
+                           value={this.state.title}
+                           name="title"
+                           type="text"
+                           onChange={this.handleChange}
+                        />
+                     </label>
+                  </p>
+                  <p>
+                     <label htmlFor="image">
+                        Image
+                        <input
+                           value={this.state.image}
+                           name="image"
+                           type="text"
+                           onChange={this.handleChange}
+                        />
+                     </label>
+                  </p>
+                  <p>
+                     <label htmlFor="link">
+                        Link
+                        <input
+                           value={this.state.link}
+                           name="link"
+                           type="text"
+                           onChange={this.handleChange}
+                        />
+                     </label>
+                  </p>
+                  <input type="submit" value="Submit" />
+               </form>
             </Modal>
          </div>
       );
